query: move QueryEvalContext and QueryNodeRef into a dedicated c_wrappers crate
