+++ src/redisearch_rs/c_wrappers/query/src/query_eval_ctx.rs
