+++ src/redisearch_rs/query_eval/tests/integration/main.rs
@@ -11,6 +11,4 @@
 redis_mock::mock_or_stub_missing_redis_c_symbols!();
 extern crate redisearch_rs;
 
-mod query_eval_ctx;
-mod query_node_ref;
 mod string_utils;
